pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "metarep"
version = "0.1.0"
description = "In-Context Learning as Meta-Representation Alignment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "wandb>=0.17",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

